+++ test/fixtures/get-pages-data-unprocessed/is-not.js
@@ -1,0 +1,5 @@
+/*---
+name: Is NOT a page
+---*/
+
+// Some content.