+++ test/fixtures/get-pages-data-unprocessed/is.js
@@ -1,0 +1,5 @@
+/*---
+name: Is a page
+---*/
+
+// Some content.